make the direct calls visible to the PE, enabling the inlining
