+++ src/som/interpreter/Method.java
@@ -1,5 +1,6 @@
 package som.interpreter;
 
+import com.oracle.truffle.api.CompilerAsserts;
 import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
 import com.oracle.truffle.api.TruffleLanguage;
@@ -7,9 +8,13 @@ import com.oracle.truffle.api.frame.FrameDescriptor;
 import com.oracle.truffle.api.frame.FrameSlot;
 import com.oracle.truffle.api.frame.FrameSlotTypeException;
 import com.oracle.truffle.api.frame.VirtualFrame;
+import com.oracle.truffle.api.nodes.DirectCallNode;
+import com.oracle.truffle.api.profiles.ValueProfile;
 
 import som.compiler.ProgramDefinitionError;
 import som.vmobjects.SAbstractObject;
+import som.vmobjects.SClass;
+import som.vmobjects.SInvokable;
 import som.vmobjects.SMethod;
 
 
@@ -21,6 +26,12 @@ public final class Method extends Invokable {
   public final FrameSlot                         stackPointerSlot;
   public final FrameSlot                         frameOnStackMarkerSlot;
 
+  private final @Children DirectCallNode[] inlineCacheDirectCallNodes;
+
+  private final @CompilationFinal(dimensions = 1) SClass[]       inlineCacheClass;
+  private final @CompilationFinal(dimensions = 1) SInvokable[]   inlineCacheInvokable;
+  private final @CompilationFinal(dimensions = 1) ValueProfile[] receiverProfiles;
+
   public Method(final TruffleLanguage<?> language, final int numberOfBytecodes,
       final SMethod method, final FrameDescriptor frameDescriptor,
       final FrameSlot executionStackSlot, final FrameSlot stackPointerSlot,
@@ -31,6 +42,10 @@ public final class Method extends Invokable {
     this.executionStackSlot = executionStackSlot;
     this.stackPointerSlot = stackPointerSlot;
     this.frameOnStackMarkerSlot = onStackSlot;
+    inlineCacheClass = new SClass[numberOfBytecodes];
+    inlineCacheInvokable = new SInvokable[numberOfBytecodes];
+    inlineCacheDirectCallNodes = new DirectCallNode[numberOfBytecodes];
+    receiverProfiles = new ValueProfile[numberOfBytecodes];
   }
 
   @Override
@@ -59,6 +74,43 @@ public final class Method extends Invokable {
         System.exit(1);
       } finally {
         marker.frameNoLongerOnStack();
+      }
+    }
+  }
+
+  public SClass getInlineCacheClass(final int bytecodeIndex) {
+    return inlineCacheClass[bytecodeIndex];
+  }
+
+  public SInvokable getInlineCacheInvokable(final int bytecodeIndex) {
+    return inlineCacheInvokable[bytecodeIndex];
+  }
+
+  public DirectCallNode getInlineCacheDirectCallNode(final int bytecodeIndex) {
+    CompilerAsserts.partialEvaluationConstant(bytecodeIndex);
+    return inlineCacheDirectCallNodes[bytecodeIndex];
+  }
+
+  public void setReceiverProfile(final int bytecodeIndex,
+      final ValueProfile valueProfile) {
+    receiverProfiles[bytecodeIndex] = valueProfile;
+  }
+
+  public ValueProfile getReceiverProfile(final int bytecodeIndex) {
+    return receiverProfiles[bytecodeIndex];
+  }
+
+  public void setInlineCache(final int bytecodeIndex, final SClass receiverClass,
+      final SInvokable invokable) {
+    CompilerDirectives.transferToInterpreterAndInvalidate();
+    inlineCacheClass[bytecodeIndex] = receiverClass;
+    inlineCacheInvokable[bytecodeIndex] = invokable;
+    if (invokable != null) {
+      if (invokable.isPrimitive()) {
+        inlineCacheDirectCallNodes[bytecodeIndex] = null;
+      } else {
+        inlineCacheDirectCallNodes[bytecodeIndex] =
+            DirectCallNode.create(((SMethod) invokable).getCallTarget());
       }
     }
   }
