+++ src/som/vmobjects/SMethod.java
@@ -26,14 +26,16 @@ package som.vmobjects;
 
 import java.util.List;
 
-import com.oracle.truffle.api.*;
+import com.oracle.truffle.api.CallTarget;
+import com.oracle.truffle.api.CompilerAsserts;
 import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
+import com.oracle.truffle.api.Truffle;
+import com.oracle.truffle.api.TruffleLanguage;
 import com.oracle.truffle.api.frame.FrameSlot;
 import com.oracle.truffle.api.frame.FrameSlotTypeException;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.nodes.DirectCallNode;
 import com.oracle.truffle.api.nodes.IndirectCallNode;
-import com.oracle.truffle.api.profiles.ValueProfile;
 
 import som.GraalSOMLanguage;
 import som.interpreter.Interpreter;
@@ -44,10 +46,6 @@ import som.vm.Universe;
 
 public class SMethod extends SAbstractObject implements SInvokable {
 
-  private final @CompilationFinal(dimensions = 1) SClass[]          inlineCacheClass;
-  private final @CompilationFinal(dimensions = 1) SInvokable[]      inlineCacheInvokable;
-  private final @CompilationFinal(dimensions = 1) DirectCallNode[]  inlineCacheDirectCallNodes;
-  private final @CompilationFinal(dimensions = 1) ValueProfile[]    receiverProfiles;
   private final @CompilationFinal(dimensions = 1) SAbstractObject[] literals;
 
   private final Method     method;
@@ -64,10 +62,6 @@ public class SMethod extends SAbstractObject implements SInvokable {
       final TruffleLanguage<?> language, final int contextLevel) {
     this.signature = signature;
     this.numberOfLocals = numberOfLocals;
-    inlineCacheClass = new SClass[numberOfBytecodes];
-    inlineCacheInvokable = new SInvokable[numberOfBytecodes];
-    inlineCacheDirectCallNodes = new DirectCallNode[numberOfBytecodes];
-    receiverProfiles = new ValueProfile[numberOfBytecodes];
     maximumNumberOfStackElements = maxNumStackElements;
     this.literals = new SAbstractObject[numberOfLiterals];
     this.contextLevel = contextLevel;
@@ -190,43 +184,6 @@ public class SMethod extends SAbstractObject implements SInvokable {
   public String toString() {
     return "Method(" + getHolder().getName().getEmbeddedString() + ">>"
         + getSignature().toString() + ")";
-  }
-
-  public SClass getInlineCacheClass(final int bytecodeIndex) {
-    return inlineCacheClass[bytecodeIndex];
-  }
-
-  public SInvokable getInlineCacheInvokable(final int bytecodeIndex) {
-    return inlineCacheInvokable[bytecodeIndex];
-  }
-
-  public DirectCallNode getInlineCacheDirectCallNode(final int bytecodeIndex) {
-    CompilerAsserts.partialEvaluationConstant(bytecodeIndex);
-    return inlineCacheDirectCallNodes[bytecodeIndex];
-  }
-
-  public void setReceiverProfile(final int bytecodeIndex,
-      final ValueProfile valueProfile) {
-    receiverProfiles[bytecodeIndex] = valueProfile;
-  }
-
-  public ValueProfile getReceiverProfile(final int bytecodeIndex) {
-    return receiverProfiles[bytecodeIndex];
-  }
-
-  public void setInlineCache(final int bytecodeIndex, final SClass receiverClass,
-      final SInvokable invokable) {
-    CompilerDirectives.transferToInterpreterAndInvalidate();
-    inlineCacheClass[bytecodeIndex] = receiverClass;
-    inlineCacheInvokable[bytecodeIndex] = invokable;
-    if (invokable != null) {
-      if (invokable.isPrimitive()) {
-        inlineCacheDirectCallNodes[bytecodeIndex] = null;
-      } else {
-        inlineCacheDirectCallNodes[bytecodeIndex] =
-            DirectCallNode.create(((SMethod) invokable).getCallTarget());
-      }
-    }
   }
 
   public final FrameSlot getStackSlot() {
